Loess interpolation + DifferentialFitnessFunction

+++ gp_prototype/src/main/java/prototype/evolution/fitness/AbsoluteErrorFitnessFunction.java
@@ -51,7 +51,7 @@ public class AbsoluteErrorFitnessFunction extends GPFitnessFunction {
             populateVariableValues(i, variablesValues);
 
             double chromosomeValueAtPointI = chromosomeAsFunction.evaluate();
-            double dataPoint = dataContainer.getValue(chromosomeVariableName, i).doubleValue();
+            double dataPoint = dataContainer.getValue(chromosomeVariableName, i);
 
             chromosomeError += Math.abs(dataPoint - chromosomeValueAtPointI);
         }
